v23tests: Expose the context to tests.
v23tests currently requires that all tests start a subprocess
to send an rpc to any other subprocess-- the parent process
can not send vanadium rpcs to its subprocsses because it does
not have the right blessings.  This change exposes the context
that has those right blessings to the test/parent process.

This change is necessary for me to test killing syncbase while
the client is in the middle of a Batch.  If the client is the
parent process, it is easy to add a kill() at a specific point.
If the client is in a subprocess, it is much harder to make
that kill() happen at the right time.

Also:
- when starting a mount table, add it to the v23tests context
  so the parent process can use it.
- remove the Principal() method.  It wasn't used anywhere, and
  now that there is a way to get the context, there is a more
  canonical way to get the princicpal:
    v23.GetPrincipal(t.Context()).

Change-Id: I57992cd1e68a28b6bbc962fdd10d1cd61b95a767