v.io/x/ref/services/syncbase/server: export types ServiceData, AppData, DbInfo
This change exports the types ServiceData, AppData, DbInfo
from v.io/x/ref/services/syncbase/server/types.vdl
That is, it captializes the names.

The purpose is to allow these types to be used in testers and checkers,
and in particular in a store consistency checker, which needs to read
all the metadata in the store.

Change-Id: Ic6b58f120cce016c780c8ee7f15dba9cc07ffec9