vdl: Make VDLRootDir stat the directory and report errors
This is used in conjunction with the recent change v.io/c/18066
that removed -builtin_vdlroot, to ensure that the vdl tool will
be able to find its vdlroot even if JIRI_ROOT or VDLROOT are set
to an invalid directory.

Note that in most cases the vdl tool should always use the
built-in vdlroot; the only reason we need the VDLROOT / JIRI_ROOT
fallback is when updating the vdlroot directory itself.

This change is necessary to make some of our tests pass, after
we've changed them to remove jiritest.NewX_DeprecatedEnv.  That's
coming in a separate CL.

Change-Id: Icad34c96c83d8bf6c4e5934ee883c8b346d75e36