ref: Add ProxyStatus to new RPC system.
ProxyStatus will just have the name and error (if any) of the proxy.
Endpoints will still have to be gotten from flow.Manager.ListeningEndpoints.

The Proxy now returns errors to rejected connecting servers. The errors are
introduce in RPCVersion13 in the ProxyResponse message.
To propagate errors to the server I had to make a slight modification to the
ProxyListen API.

The change also removes some tests that were redundant but needed during
the transition.

MultiPart: 2/2

Change-Id: I5ff989457de8ba8453dc03a3099d69f328f435c0