syncbase: Handle corrupt leveldb.
If a leveldb is corrupt, syncbase will move it aside and return an error.  It
would be nice to recreate the database so the app would be ready to go next
time. Because the database stores its own ACL, syncbase can't just recreate
the database.  Apps need to create their app/db/table hierarchy on startup
anyway, so it isn't a big deal to just have the app recreate the database.

- If an app's db is corrupt, remove that DB from the top-level leveldb.
- If the top-level leveldb is corrupt, move aside all of the app databases.

MultiPart: 1/2

Change-Id: I1aa7e3cbf627ebf7c4ccd6692f4b91e9be426e0d