flow/manager: Fix caching of connections.
Connections should not be keyed by blessingNames, instead the
blessingNames are compared upon retrieval from the cache (only
if the conn is not to a proxy).

This ensures that connections to proxies from servers aren't replaced
from the cache when a server makes a connection through the proxy
(acting as a client). The blessings of the end server were incorrectly
being used in the caching of connection to the proxy.

Change-Id: I4a30f3610820be6dd704095653612481bebff627