services/agent/pod_agentd: Make socket 666
Change the permissions on the socket to be world accessible. This allows
the applications inside the containers to run with a non-root UID.

Note that the parent directory of the socket is only mounted in
containers that should have access to it. So, this does not give
anything access to the principal to anything new.

Change-Id: I5ef033fdb738ad4c007a61d9a9a279a1abbb6b4a