flow: Fix two bugs with the flow control counters:
1) sometimes we doubled released counts when a flow closed if a
release message arrived simultaneously.
2) If we run out of counters on the first write to a flow we would
deactivate it before we sent the OpenFlow message.  Then nothing
would reactiveate the flow since we only reactivate when we
receive released counters for that flow and since the other side
doesn't know about it, we never sent that release.  Now we always
send the open flow, even if we can't send any payload bytes.

Change-Id: Ib40cf1b416586c08b07e8c922b71d8014a087222