TBR syncbase: increase healthInfoTimeOut
This fixes a peer selection slowdown problem introduced by
v.io/c/18414:
- v.io/c/18414 was needed because a 200ms connection timeout
  caused sync to never succeed for Kishore's GCE -> proxied
  phone configuration.
- This CL is needed because if healthInfoTimeOut is less
  than connectionTimeOut, we can end up in a similar place
  to where we started, where an unresponsive peer causes
  sync to slow down.

With this CL, we expect the following behavior: when a
device becomes unresponsive, its peers may have up to a 4s
"hiccup" in sync. But after this hiccup, sync should again
be fast. This isn't ideal, and we know of various ways to
improve things further, but it solves the immediate problem.

Ali patched this CL to test it and confirmed that the new
behavior is as described above.

Change-Id: I5e8f8dd82876fe121e7cd00c11b1157143bf719d