Add support for unknown values during conflict resolution.
The conflict resolution workflow can pull local objects which
were not under conflict into a conflict batch that is
sent to the App. Since the remote value of the object is not
known, the conflict needs to provide the App with enough
information to distinguish between a deleted version and a
version that is unknown.

MultiPart: 2/2

Change-Id: Ia06c82f98313918556395eafc4052e1b6ef32e99