./cmd/vrun: first step in deprecating vrun
Prevent users from running vrun anymore and point them to vbecome
instead; but offer an escape hatch if they really need to.

Background: vrun is being deprecated in favor of vbecome. Vrun's woes
stem from its usage of exec to run the payload command: this prevents
the tool from cleaning up the socket file created for the command to
communicate with the agent; and it also runs into race issues with the
go runtime's pthread creation

Change-Id: I17f106d7fd637999585d3391002a87cf3401f2bb