services/debug/debug: "Browse" interface for debugging.
The debug command-line tool presents a wealth of information
useful when inspecting remote servers, however, the command-line
nature of it can be a bit daunting. This patch introduces a
"debug browse" command that starts up a local webserver
that provides a web-interface to browse the information available
via the command-line tool.

Change-Id: I769d41152f87a1130efe46698561c1150b471abf