query: implement execution of delete statements.
Parsing and semantic checking previously submitted.  This CL
adds execution of delete statements and includes query
package tests.

Delete is completely wired for the db test implementation in
the query package tests, but not yet wired into syncbase; hence,
syncbase will return a NotWritable error if a delete is
attempted.

Discovery returns the same error (and this won't change as
discovery is readonly).

MultiPart: 2/2
Change-Id: I36a7ac53c1058bf920a68e8c2dc6209c6d28aad4