ref: remaining syncbase vclock featuretests
Along the way, this CL also:
- Does some refactoring to share a common helper for
initializing a set of Syncbase instances
- Fixes logging prefix in vclockd to be "vclockd", not
"vclock" (artifact of previous code move)
- Tweaks "MaxNumHops" constant to mean, num hops from self
to NTP source. I found the previous definition (num hops
from peer to NTP source) confusing, and I suspect it was
not a deliberate choice to define it that way vs. "from
self".
- Renames CR tests to all start with "TestV23CR" so that
it's easy to run CR featuretests but not other
featuretests. This is in line with our convention in other
files, and was helpful for testing this change.
- Sprinkled in a few TODOs where appropriate
- Possibly a few other minor tweaks

MultiPart: 2/2

Change-Id: Ic4d3089676e1f4218ceac11f5c871fc617ebc75b