syncbase: Restartable database destroy. Destroy databases on app destroy.
Database store paths were changed to include a uuid, allowing stores
to be garbage collected independently of deleting/recreating databases.
Stores not referenced by databases are atomically written to a
separate log, ensuring garbage collection after failure or crash.
Databases in an app are atomically destroyed when the app is destroyed.

MultiPart: 1/2
Change-Id: I6c084d4b9523003142bc3ef6d3373356874ed6a8