"x/ref": Update RPC benchmarks
This CL updates the private mutual authentication benchmarks
to account for multiple blessing patterns in the server's
authorization policy. At this time, there are no noticeable
changes in the connection establishment time with increasing
the number of blessing patterns. This is because each ciphertext
sent by the server identifies the blessing pattern it is encrypted
for, allowing clients to choose the ciphertext whose pattern is
matched by their blessings. Thus, the number of decryption operations
performed by the client stay the same.

Change-Id: I877e3d65112c6de13e88e268133ed84c95b3f06c