query: wire up query's delete to syncbase.
Previous CLs imlemented delete in the query package.  This change
wires it up to syncbase.  The complicating factor is that, in
the non-batch case, we don't know if we should create a
snapshot (readonly) or a transaction (read/write) when
query/engine.Exec is called.  It is only when the GetTable
function is called (on the database interface) that syncbase
knows whether it should create a snapshot or a transaction.

MultiPart: 1/2
Change-Id: Idd5637806da3ec547ed59dbe7e3a546c582efb2a