ref: Set root cancel on context of encapsulated flows in Conn.
Proxied RPCs were failing in the following scenario:
A starts proxied RPC 1 to B with ctx, cancel.
A starts proxied RPC 2 to B.
RPC 1 completes and A calls cancel.
RPC 2 fails.

This was happening because cancel was closing the encapsulated proxy
flow. The solution is to use SetDeadlineContext to set the ctx on
the flow the RootCancel context used in the setup of the encapsulating
Conn.

Change-Id: I56455d52ee6dcd579fa1e92c00ce63ddb49e18f9