rpc: Fix bug in discharge preparation.
Resolves vanadium/issues#1049
The issue was that we were incorrectly equating a list with an empty
discharge to a map with a single valid discharge because we'd compare
the zero value from the map lookup. I apologize for this shoddy
description. The added unittest would fail without the corresponding
fixes with:
	full_test.go:854: test.test:"mountpoint/name".ResolveStep:
	Access denied: [] [{"test-blessing:client":
	test.test:"mountpoint/name".ResolveStep: caveat validation
	failed: test.test:"mountpoint/name".ResolveStep:missing
	discharge for third party
	caveat(id=eSO76jSotl6b3Jpe2NdZ4BfmjFV1P6tWjBDxnRhv+sY=)}]

(Bug introduced in https://vanadium-review.googlesource.com/#/c/18560/)

Also fix a bug that would manifest if the discharge fetching failed.

Change-Id: I50e366065c5779563e5c4ef7de73c1a320e2ccca