ref: Added a visualization page for vtrace.
The current implementation uses colored divs to represent spans in the
trace.  The width of the div corresponds to the time that the span
took.  The name of the span is available on hover.  There the
beginnings of a tree structure on the left side of the span
visualizations to allow collapsing spans and see the parent heirarchy of
the spans.

MultiPart: 2/2
Change-Id: I1250c7f80c728f177e8e0ddf081634ed2f12ba5b