services/agent/vbecome: Add explicit timeout
Add an explicit timeout to bypass the default 30-sec name resolution
timeout.

If we use a /host:port/role address to talk to roled, the client first
tries to resolve the name (with a ResolveStep RPC to /host:port). If
this RPC times out, the SeekBlessings RPC fails.

This is a problem because roled's private key is stored remotely, and
the first RPC sometimes takes more than 30 seconds to complete (but
less than the new default timeout).

Change-Id: I4d39b76f3ed52f91d2b6c3a1b9eea074cf6247f7