TBR: ref: Change gosh.Cmd.StdinPipe to have unlimited-size.
It's more convenient for users if StdinPipe has an unlimited
size; otherwise they need to worry about deadlock.  You might
think we could just use our buffered pipe and be done with it,
but it's pretty tricky.  In particular os/exec will create its
own os.Pipe and a copier goroutine if we don't pass in a *os.File
for exec.Cmd.Stdin, and exec.Cmd.Wait will wait for that
goroutine to finish.

Note that we can't call exec.Cmd.StdinPipe; that ensures that
Close is only called once, but doesn't ensure Write and Close
calls are synchronous.  This makes a race possible: the user may
call Write on the StdinPipe, concurrently with the exit-waiter
goroutine calling Close on the StdinPipe when the process exits.
This race has shown up on jenkins, and is also easily
reproduced (under heavy machine load) with the new test.

To make this all work, we create our own os.Pipe that we set for
exec.Cmd.Stdin, and also create a buffered pipe, along with our
own copier goroutine.

MultiPart: 2/3

Change-Id: I4937533cb099af53748ee174327da3927f522dde