services/agent/internal: add lock that handles concurrent stale lock reclaiming.
The lockfile package defines a lock implementation that recovers from
stale locks (owned by dead processes), but potentially allows multiple
processes to reclaim the same stale lock.

The new lock package tries to ensure that only one process can reclaim
the stale lock.  It does so by employing a hierarchy of lock files (see
package comment for description of the algorithm).

A less significant difference is that the new lock package associates
the lock with a directory instead of a file (more suitable for how I
intend to use it to guard the credentials dir).

Some utility code (CreatePIDFile and StillRunning) is refactored into a
new lockutil package to be shared between lockfile and lock.

Change-Id: Iac37fbe15d031abd9c4532de0be84c0c7da905b0