rpc: Ensure that typeFlows are created on the same conn as the first rpc flow.
This wasn't happening when a hostname used in an endpoint resolved to
multiple ip addresses. This is because we are relying on the cache to
return the same Conn for a dailed type flow as the first regular flow
which is not the case when the ip address used to look up into the cache
changes.

Fix:
(1) Ensure the Endpoint returned by flow.RemoteEndpoint includes the rid
    of the remote end (if the remote end is a server).
(2) Use flow.RemoteEndpoint as the endpoint argument when dialing the
    type flow.

Additionally, this change also caches connections based on the unresolved
network and address of the dial call.

Change-Id: Idde4529d9cac06eaafab4f3b771e5cf06fde744f