v.io/x/ref/services/syncbase/localblobstore: BlobWriter.AppendFragment -> AppendBytes
This change renames BlobWriter.AppendFragment to BlobWriter.AppendBytes,
and modifies the code to buffer up to some threshold (currently 1MByte)
before creating a fragment (a file on disc).

This largely does not affect usage, just makes performance more independent of
how the call is used.

The only semantic changes is that data written to a blob is not necessarily
committed to disc until the blobwriter is closed.  No uses relied on the
previous behaviour.

This change also fixes an error in the old localblobstore_transer_test.go
which handled a transfer incorrectly if the receiver had chunks
that were not a simple prefix of the blob being transferred.
The code has been changed from a Go Example to three Tests
in order to test this case, plus the other possible cases.

Change-Id: I97c39cccc6c4edc17684155a848d156e58fc4ba6