raft: add a Sync() call.
It returns when the member catches up
to the current commit point.  Thus causal reads can be performed
from follower members by preceding the read with a Sync.

To avoid an RPC per request, I added a syncLoop go routine
that performs the sync whenever requested and will batch up
any requests that occur while the last sync is in progress.
It does mean that there may be back to back syncs to the
leader under load.  I could always throw in a ms sleep
if need be.

Change-Id: Id8d901f3714d6451f0cde92f7e26e88b3fa82808