raft: Fix jenkins failures.
- whenever an rpc sends or returns a term larger that our
currentTerm, set currentTerm to that term and become a follower.
This is explicitly stated in tiny little text in Figure 2 of the
paper.  What a fool I was to miss it.

- clean up test timeouts so that we don't give up even if multiple
servers start overlappping elections several times in a row.
It can still hapen but is very unlikely.

- make the candidate timeout longer than the follower timeout.
This is not necessary but helps elections terminate quicker
since 'bad' candidates (i.e. ones that have incomplete logs)
will back off after failing to win an election giving other
followers a chance to start their own election.

Change-Id: Ib5a946cb50e1d72bc160d8b33788b13742c77f76