TBR: flow/conn: Properly initialize RTT upon conn creation.
We were not setting the first RTT value for newly created connections.

With this change we measure the first dialer RTT from the sending of
the Setup message to the receipt of the acceptors auth message.
And we measure the acceptors RTT from the sending of the auth message
to the receipt of the dialers auth message.

Unfortunately, the acceptor RTT is an overestimate because the it
includes the dialer's time spent verifying the acceptors certificates.
I think this wont really be a problem because users of RTT will usually
have a short channel timeout which should refresh the RTT time to a
more accurate time soon. In the syncbase case, it will get refreshed in
2 seconds.

If this proves to be a problem for accepted conns we can do a more
involved fix.

Change-Id: I03d82b4ecf28bdda2d9aa48efe5f44bc7231b823