ref/services/identity: Add a secondary index to the auditor table
The only SELECT query we do is using a WHERE on the Email and and
Timestamp. This secondary index this CL adds will make that efficient.

The change from NVARCHAR(256) to VARBINARY(256) for the Email column
is require because a unicode character can take up to 4 bytes and that
MySQL doesn't index columns that are larger than 767 bytes. The max
side in the prod instance is currently 53 bytes.

Note: making the (Email, Timestamp) primary key doesn't work because
Timestamp is using 1-second granularity and that generates quite a bit
of duplicates (>30K in our prod instance).

Change-Id: I86b64551f7a94a3c6d47cc4528ae34b85925d545