services/device/deviced/internal/impl: remove race from reaper test
TestDaemonRestart suffered from the following race: the app gets
restarted, but before the device manager updates the state to 'running',
the test queries for the state and finds it in state 'launching'. To
fix, we wait for the state to update rather than querying for it once.

To make the test more robust, avoid using a second app instance just to
trigger the daemon poll -- it's too brittle, and interferes with the
ping channel.  Just rely the periodic poll.

Finally, make the restarts happen more than once, to test that after a
restart, the app is still properly monitored by the reaper.

Change-Id: Iec3c0a0858b4e8e6e31055e8d64941a68ba7d30b