services/debug/debug: Support multiple names
With this change "debug browse" accepts multiple object names and uses
the first one it can connect to.

This was motivated by my changes to the Android APIs to add a "remote
debugger" call which emails the command-line that another user can
employ to debug the Vanadium services running on the phone. This debug
server has a local network address and also a proxied address (which it
mounts in the namespace). With this change a single "debug browse"
invocation has a higher chance of successfully connecting to the debug
server in the Android application.

Change-Id: I5ce8058ff349d5a36af75a87f6f8312f7da1f610