services/device/deviced/internal/impl: simplify reaper and remove flakiness
The reaper initialization logic was too complicated and race-prone (apps
found to be dead when the reaper came up had two paths by which they
could be restarted --
https://github.com/vanadium/issues/issues/1018). This CL removes the
explicit restarts that were happening at the end of newReaper(), and
relies exclusively on the polling loop in processStatusPolling() to
discover restart candidates and restart them.

Another unrelated fix is to make the shutdown of the
processStatusPolling loop thread-safe w.r.t. reaper commands (the old
code could result in sending on a closed channel, which panics).

While at it, simplify the code by removing the code path that does
reconciliation via the app cycle manager, since it's not used.

Unrelated to the above, apply the sample fix to
TestReapRestartsDaemonMode as in
https://vanadium-review.googlesource.com/#/c/20162/ and re-enable it.

Change-Id: I1b15b13fd6c13c67e90890039d3cfb2883e79708