"x/ref/services": IBE Root service
This CL defines the interface and implementation for a service
for exchanging blessings (from a certain identity-provider) for
an IBE private key that can be used for decrypting any message
encrypted for a pattern matched by the blessing. The private keys
are generated using an IBE Master object held by the service. The
service is meant to be run alongside the identity provider for
the blessings.

Change-Id: Ie55781d2989862e88780cceaf517002fe891abf7