syncbase: Watchable.Store refactoring
The main motivation of this change is to split the watchable store in
a generic part and syncbase-specific part. The generic part is not
fully generic yet (PutOp and DeleteOp still carry the PermKey and
PermVersion, FromSync is still in LogEntry). A few other smaller
cleanups and refactoring are also included.

The stack of changes that are merged in here is the following:

  http://v.io/c/20644 syncbase: Remove the *watchable.Store from mockService
  http://v.io/c/20645 syncbase: Remove the GetStoreTime function
  http://v.io/c/20646 syncbase: Rename the VClock to Clock and move to common/
  http://v.io/c/20647 syncbase: Move syncbase/server/watchable to syncbase/watchable
  http://v.io/c/20648 syncbase: Address some of the comments to v.io/c/20646/5
  http://v.io/c/20723 syncbase: Split the syncbase/server/util
  http://v.io/c/20724 syncbase: Remove the watchable.Store interface
  http://v.io/c/20725 syncbase: Make watchable.transaction public
  http://v.io/c/20820 syncbase: Switch from store.Store to watchable.Store
  http://v.io/c/20821 syncbase: Watchable.Store refactoring

MultiPart: 1/2
Change-Id: I5e4fa0816624e5080e61731ac1c9106afb5621ed