x/ref/services/syncbase: Add CPU Profile flag to syncbased
Add CPU Profile flag to better understand where our time is spent in
syncbased. This flag requires the import of runtime/pprof.

Original Size: 23,790,770 bytes
New Size: 23,795,241 bytes

The delta is only about 4500 bytes, which is about 0.02% of the total.

CPU Profile appears to randomly sample the program in a goroutine, so
that may explain why it doesn't add much to the binary.

The test utility was also modified so that when the
-v23test-syncbase-debug-args flag is set, we will also do CPU profiling.

Note: It may be better to make cpu profiling only turn on if it is
mentioned in the syncbaseDebugArgs string.

Change-Id: Iab673c6e8368946e4e634f650b9afffe02daf4f1