services/internal/logreader: Android support
This patch implements the LogFileServer interface for Android.  With
this, servers in Android provide the same quality of information as
servers on Linux/Darwin machines do. In particular, "debug logs read"
and "debug browse" let you inspect application logs.

(On Android, all logging is deferred to the Android logger instead of
files, so this implementation uses "logcat" instead of inspecting
files to serve content)

Change-Id: Ib85ef909b30c3a8ae32aa0a206fa3f7f31262eab