lib/discovery: Handle lack of BLE initialization in Android gracefully.
https://vanadium-review.googlesource.com/#/c/20423/1/lib/discovery/factory/factory.go
changed behavior to (correctly) use all registered protocols for
discovery. However, on Android it had the side effect of discovery
initialization failing if SetPluginFactory("ble", ...) was not called
by the Java code.

As a result, pure Go code that used discovery (such as benchmarks in
v.io/v23/syncbase/nosql run through bendroid) started failing with
"ble factory not initalized".

With this commit, instead of failing, those tests will simply
not try to use BLE as a discovery protocol.

Change-Id: I5019714f5cef9767c77afdc7552468d0b59fc277