device/deviced/internal/impl: fix a couple races
1. the grace period we give to the app to exit after forcibly
terminating it is too short (0.5 s) and occasionally we fail to observe
the process' exit before the grace period expires. I've seen that a few
times on Jenkins. Boost the grace to 5 s.

2. if the app is being explicitly Killed or Run, we should avoid having
the reaper try to restart it. The solution is to have markNotRunning
return an error if the app is not in state Running prior to attempting a
restart (it would be in state Dying if a Kill is underway or Launching
if a Run is underway).

Change-Id: Id76b1b98d53a704bf7368d692929c6b5e0e62121