vdl: Always use the built-in vdlroot standard packages.
Before this change, we had vdlroot standard packages built-in to
the vdl binary.  This support was used in the java/gradle world,
but wasn't used elsewhere.  We'd simply forgotten to complete the
migration.

We also had some technical issues; e.g. we'd need to move the
built-in vdlroot from the vdl binary into the vdl/build package,
so that all users would be able to call build.TransitivePackages,
even if they didn't have the vdl sources.  But that would be
messy, since the previous compiled-in tarball approach required
that we untar into a temp directory, and cleanup is painful.

After this change, we use go-bindata to archive the vdlroot
packages into the vdl/build package, which we can access directly
from memory.  This sidesteps the tmpdir cleanup issue, and is
also generally cleaner.  As a result we can also remove the
long-standing hack of falling back on JIRI_ROOT to find the
sources if VDLROOT isn't set.

In the new world, the vdl compiler will use the built-in vdlroot
sources by default.  You can override this by setting VDLROOT;
that's still needed in order for us to be able to update the
vdlroot sources.  This is better behavior.  We shouldn't have
required that the vdlroot sources were available for end users to
use the vdl tool in the first place; that was meant as a
temporary hack.

Fixes v.io/i/1228

Change-Id: If71e1663221d508992030a032737c6ef138ff72c