rpc: Fix server closed function.
The server closed function returned the s.ctx.Done channel which was
closed when we give up waiting for active requests in the server to
finish. This means that goroutines could still be running when
server.Closed was closed. I believe this is why we are still running
into issues where we have panics during runtime initialization.

This change fixes this by closing a different channel when goroutines
from the publisher and flow manager MUST have completed.

Fixes vanadium/build#32

Change-Id: Ia79a57beb1d0f041515d5a7b4b6b349c68de1178