services/internal/dirprinter: dump directory contents when test fails
This CL introduces the dirprinter package, which contains utilities for
dumping the contents and structure of a directory tree.  We use it to
get better insight into test failures where significant clues may lie in
the directories left behind (e.g. devicemanager).  This should be
particularly useful now that we cannot get to the directories after the
test finishes on Kubernetes.

Change-Id: Ica531b3a3dce0edc4197e9c25a0cef6e7023dcaf