test/expect: call FailNow by default upon error but make behavior configurable
It's rare that we want to continue when expect encounters an error.  The
cost of not failing fast is 10 minute test timeouts.  This change makes
the behavior as to whether to fail fast or not configurable on the
Session; for Sessions created via the Shell, respect the Shell's
ContinueOnError setting.

Change-Id: I101fd27c6511c88589002a1891a78d095e34c223