vine: Change to adjacency matrix input.
The api for SetBehaviors takes a map keys by both the dialing and accepting
ends of a connection.

Now, reachability is enforced on either/both the accepting and dialing sides
of a connection.

Additionally, vine.WithLocalTag sets the tag for a context that will be used
when dialing and listening on addresses, so we no longer need to manually specify
the tag in the vine address, instead the Listen function does it for us.

Change-Id: I08c87be5d53cde2f7967b05727347cde4cfaeac3