services/cluster/vkube: update with persistent disk
Rolling updates with persistent disks are problematic because the new
Pod can't come up until the persistent disk is released. The best thing
we can do in this case is to bring up the new RC, delete the old one,
and hope for the best.

Also, treat updates with a RC name that already exists as a no op, i.e.
assume that the update was already applied.

Change-Id: I74d6b2d70af81a32190c2666ff498b0d66cdb004