service/cluster/vkube: add support for Deployment
Deployment is a new, better way, to run replicated services on
Kubernetes. http://kubernetes.io/docs/user-guide/deployments/

This change adds tentative support for Deployment. It is not yet
available on GKE. So, this change is barely tested. I just made sure
that the support for Replication Controller still works.

Change-Id: Ia1a0479aa1f0b1be16d7cbb3b3b81fe03b773155