services/device/deviced/internal: fix testDeviceManager to use test credentials
We generate separate credentials for testing a new version of device
manager. However, we didn't actually make the test device manager use
the new credentials when not using the agent: the existing device
manager credentials were set in the deviced.sh script, and took
precedence over the setting of the new credentials in the environment of
the command. The fix is to use the config override and set
v23.credentials to the new credentials, thereby overriding whatever was
set through the environment in deviced.sh.

Fixing this uncovered a missed replacement of / to : in the blessings
separator. This was not a problem up to now since the test device
manager used the same credentials as the existing device manager and
hence didn't really need to run in test mode and bypass the signature
checks on acls.

Change-Id: I33d631998949c7b6467f76808abdf1eec4f66b08