services/agent/internal/ipc: Wait for agent
If the agent is not ready yet, wait for it.

This happens on kubernetes where we have two containers: one for the
application, one for the pod-agent. It pod-agent isn't ready first, the
application fails and gets restarted. This typically happens once or
twice every time a Pod is started.

With this change, the application will wait for the agent to be ready
instead of failing.

Change-Id: I8589d88cb606ea7a3c1773764a27074367f1790a