services/ben/benarchd: Dynamic template/asset loading.
This patch re-organizes usage of templates and static assets used for
the web-interface provided by benarchd so that they can be dynamically
loaded from the filesystem.

In "production", benarchd remains a standalone binary with all the
templates and assets compiled into the binary. However, when the
--assets flag is specified, then it re-loads (and re-parses) files (and
templates) from the filesystem - enabling faster development iteration
on this web interface.

To compile in those same assets, we rely on "go generate" and
https://github.com/jteeuwen/go-bindata

Change-Id: Ida1d3e427127bb0b6ae8b4a5ca238db9a6a041af