services/agent: expose the socket connection timeout in the API
The newly introduced connection retry logic (with a 60 second timeout)
caused the TestV23AgentPrincipal integration test in
services/agent/agentlib to balloon from ~8 seconds to 12 minutes.  This
is because the LoadPrincipal logic in agentlib tries to connect to the
socket (repeatedly) before potentially launching an agent or loading the
principal locally.  If each attempt takes 60 seconds, we're not getting
anywhere.

To make LoadPrincipal work again in a sensible amount of time, this CL
exposes the timeout as a knob in the NewAgentPrincipal API.  As luck
would have it, NewAgentPrincipal was available, so the CL introduces
NewAgentPrincipal taking a path and a timeout.  The existing
NewAgentPrincipalX is unchanged, and uses the 60 second timeout
introduced by Robin.  I plan to deprecate NewAgentPrincipalX in due
time.

While at it, rewrite the ipc.Connect logic to not overshoot the timeout
value by up to a second, and to avoid the use of breaks and continues.

Note, I was planning to do something like this anyway in order to
implement some of the fallback logic in LoadPrincipal down the road, so
this is not purely to conditionally bypass Robin's change.

Change-Id: Ia326d8ebba1cb917038cc9d4dc1d9541df4f1ac2