Model definitions and functions for longevity test control script.
Types for model Databases, Devices, Users, and Universe, and functions
to generate random sets of them.

The "top" function is GenerateUniverse, which generates a set of
databases, devices, users, and a topology on the devices.  The output of
this functions will be used to construct a network of actual running syncbase instances
with the specified connectivity and topology.

There's still a lot missing, including collections, blessings, and any
sort of dynamicism.

Change-Id: I52210be71b980f6b0be53a16b3b5c211e6e448e8