services/cluster/vkube: Add readiness check & debugging
Trying to figure out a weird test failure:
vkube> Error from server: container "tunneld" in pod
"tunneld-106443024-2dtme" is waiting to start: ContainerCreating

This happened while fetching the logs of that container after the
deployment rollout was supposedly done.

Added a readiness check to see if it makes a difference, and added extra
logging.

Change-Id: Ie7312ef0931d72b9d5a4620c55ceaacecfe5cc8b