services/cluster/vkube: Test improvements
There is a bug somewhere in kubernetes that makes 'kubectl logs'
sometimes hang forever. Added a 5-sec timeout and a retry around it as a
workaround.

Added flags to control where the docker images are pushed and how often
they are rebuilt. This allows the test to run in less than 10 minutes
with prebuilt images, instead of 30+ minutes otherwise.

Added a readiness probe to the cluster-agent container to make sure that
when 'start-cluster-agent --wait' and 'update-cluster-agent' --wait
return, the cluster agent is actually ready.

Change-Id: I60cd4c5bafa27194b9a6f828fc3c4e0cbe6f464e