services/cluster/vkube: bug fixes
Change watchDeploymentRollout() to fix a typo (metadada) and guard against
missing values better.

Change readyPods() to only report pods as ready when all their containers
are ready.

Change updateClusterAgent() to wait for the previous incarnation to be
gone before starting the new one. This was only an issue for the
integration test and will go away completely when we switch
cluster-agent to use Deployment.

Change-Id: I88db41391cc651abf299670750c816c439869b65