services/cluster/vkube: back to killing process group
It turns out we do need to kill the child's process group after all.

This is due to how exec.Cmd.Wait() works. After the process exits,
Wait() blocks until all the data from stdout and stderr has been read,
i.e. reads return an error.

In the case of 'vkube kubectl', vkube runs kubectl with Stdout and
Stderr set to vkube's own os.Stdout and os.Stderr. This means that even
if vkube is dead, Wait() will not return until kubectl is also done and
its stdout & stderr are closed.

This is a pretty annoying quirk of go. The documentation for Cmd.Wait
mentions this behavior for Stdin, but not for Stdout & Stderr.

https://golang.org/pkg/os/exec/#Cmd.Wait

But looking at the implementation, we see that Cmd.Start() creates 2
goroutines to read from the child's stdout & stderr, and Cmd.Wait()
blocks until they exit.

The easiest workaround for us is to kill the process group of the child.

Change-Id: I7e0526b9ef91eeb308d7458558b142bbecab2168